functions/process_labeling_export/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-storage
